requirements.txt
annotated-types==0.7.0
dnspython==2.7.0
paho-mqtt==2.1.0
pika==1.3.2
pydantic==2.11.3
pydantic_core==2.33.1
pymongo==4.12.0
typing-inspection==0.4.0
typing_extensions==4.13.2
